+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 # 3.6.6
 - New DZRP command CMD_INTERRUPT_ON_OFF in use for cspect. IMPORTANT: You need Cspect >= 3.0.15.2 to be compatible with DeZog.
 - Fixed issue loading .sna and .z80 files if interrupt is on and cspect is the target.
-
+- Fixes in zsim configuration:
+  - It is now possible to deselect the ulaScreen
+  - zxKeyboard is now disabled by default
 
 # 3.6.5
 - Debug command "-sjasmplus.path" added for dynamic loading of sld files during runtime (#156).